requirements.txt
-i https://pypi.org/simple
altgraph==0.17.2
astroid==2.11.7; python_full_version >= '3.6.2'
blessed==1.19.1; python_version >= '2.7'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
colorama==0.4.5; sys_platform == 'win32'
commonmark==0.9.1
coverage==6.4.4
dill==0.3.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
google-api-core==2.8.2; python_version >= '3.6'
google-api-python-client==2.57.0
google-auth==2.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4; python_version >= '3.7'
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3.5'
inquirer==2.10.0
isort==5.10.1; python_version < '4' and python_full_version >= '3.6.1'
lazy-object-proxy==1.7.1; python_version >= '3.6'
mccabe==0.7.0; python_version >= '3.6'
oauthlib==3.2.0; python_version >= '3.6'
platformdirs==2.5.2; python_version >= '3.7'
protobuf==4.21.5; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygments==2.13.0; python_version >= '3.6'
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8; python_version >= '3.7'
pylint==2.14.5
pyparsing==3.0.9; python_version >= '3.1'
python-editor==1.0.4
readchar==4.0.1; python_version >= '3.6'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==12.5.1
rsa==4.9; python_version >= '3.6'
setuptools==65.0.2; python_version >= '3.7'
shellingham==1.5.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.8.10
tomli==2.0.1; python_version < '3.11'
tomlkit==0.11.4; python_version >= '3.6' and python_version < '4'
typer[all]==0.6.1
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wcwidth==0.2.5
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
